forge fmt

+++ src/security-1/Denial.sol
@@ -2,11 +2,10 @@
 pragma solidity ^0.8.24;
 
 contract Denial {
-
     address public partner; // withdraw partner - pay the gas, split the withdraw
     address public constant owner = address(0xA9E);
-    uint timeLastWithdrawn;
-    mapping(address => uint) withdrawPartnerBalances; // keep track of partners balances
+    uint256 timeLastWithdrawn;
+    mapping(address => uint256) withdrawPartnerBalances; // keep track of partners balances
 
     function setWithdrawPartner(address _partner) public {
         partner = _partner;
@@ -14,10 +13,10 @@ contract Denial {
 
     // withdraw 1% to recipient and 1% to owner
     function withdraw() public {
-        uint amountToSend = address(this).balance / 100;
+        uint256 amountToSend = address(this).balance / 100;
         // perform a call without checking return
         // The recipient can revert, the owner will still get their share
-        partner.call{value:amountToSend}("");
+        partner.call{value: amountToSend}("");
         payable(owner).transfer(amountToSend);
         // keep the time of the withdrawal
         timeLastWithdrawn = block.timestamp;
@@ -28,12 +27,14 @@ contract Denial {
     receive() external payable {}
 
     // convenience function
-    function contractBalance() public view returns (uint) {
+    function contractBalance() public view returns (uint256) {
         return address(this).balance;
     }
 }
 
-/** CODE YOUR SOLUTION HERE */
+/**
+ * CODE YOUR SOLUTION HERE
+ */
 contract Attacker {
     Denial public victim;
